+++ plugin-maven/src/test/java/com/diffplug/spotless/maven/MavenIntegrationTest.java
@@ -102,6 +102,10 @@ public class MavenIntegrationTest extends ResourceHarness {
 		writePom(groupWithSteps("kotlin", steps));
 	}
 
+	protected void writePomWithCppSteps(String... steps) throws IOException {
+		writePom(groupWithSteps("cpp", steps));
+	}
+
 	protected void writePom(String... configuration) throws IOException {
 		writePom(null, configuration);
 	}